Organize browser tools benchmark modules
